Initial commit
Initialize this project dedicated to finding "gadgets" in the JavaScript
language itself with 24 gadgets to start. The project includes a proof
of concept for each of the 24 gadgets and describes each of them. It
also lists sections of the spec that were evaluated but deemed
unaffected.

The process for getting these 24 gadgets was entirely manual and very
time consuming.