Setup infrastructure for finding gadgets using `tc39/test362`
Initialize the `tc39` directory with the `tc39/test362` repository as a
git submodule. Furthermore, this directory contains several files that
aid in using this test suite for analysis purposes. In particular, it
defines a Node.js+npm project that uses 3 direct dependencies to define
a trio of Node.js script that together form the analysis. `rewriter.js`
takes care of rewriting tests in the suite for analysis purposes, then
`analyze.js` takes care of analyzing a test suite run's output and
produces a JSON file with relevant results, finally `to-sarif.js` can
convert that JSON file into a SARIF file to aid in manual review (more
detail in `tc39/README.md`). These scripts are combined and run in
series by `run.sh` which is the "entrypoint" for the analysis. A OCI
container definition is included to allow for running in the analysis in
a more controlled environment.

As of this commit, this infrastructure was used to find 4 additional
gadgets and rule out 1.