Find 6 new gadgets based on test262 suite
Use the results from the test262 to create 6 additional PoCs of gadgets
in the language.

One gadget is in the `with` statement, which is only available in non-
strict CommonJS. For this reason it's not included in the existing PoC
infrastructure which is purely ESM. Running CJS is not supported by
Deno and how it can be run differs between Node.js and Browsers. For
now, you can test it with `node pocs/with-@@unscopables.PoC.cjs` (have
not attempted it in Browsers yet).

Also improve the project infrastructure by providing make commands to
run the PoCs in a Docker container for improved reproducibility.